% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqdist.R
\name{freqdist}
\alias{freqdist}
\title{Generates a Frequency Distribution with columns for raw frequencies, percentages, and cumulative percentages.}
\usage{
freqdist(x)
}
\arguments{
\item{x}{Variable on which the frequency distribution is generated.}
}
\value{
- Returns the frequency distribution.
}
\description{
Generates a Frequency Distribution with columns for raw frequencies, percentages, and cumulative percentages.
}
\examples{
# Create Sample Data Frame

data <- c(1,2,3,4,4)
data <- as.data.frame(data)

# Run Frequency Distribution Command on Data Frame

freqdist(data)
}
\author{
Thomas Feldman and Kenneth W. Moffett
}

